feat(bet4-principle): extract principle_decide into a scoped ToolPack — executeTool switch drained to zero named tools (EP-8DC217EB BET-4)
Process-Spine-Decision: BET-4 drain-to-zero (plan §4) — behavior-identical; the kernel decision gate is the final named case, completing the drain.
BI: BI-2B7EE073
