+++ apps/web/lib/mcp/pack-registry.ts
@@ -74,6 +74,7 @@ import { adminPack } from "./packs/admin-pack";
 import { decompositionPack } from "./packs/decomposition-pack";
 import { buildVisibilityPack } from "./packs/build-visibility-pack";
 import { buildEvidenceExtraPack } from "./packs/build-evidence-extra-pack";
+import { principleDecidePack } from "./packs/principle-decide-pack";
 
 import { grokSigninPack } from "./packs/grok-signin-pack";
 import { releasePack } from "./packs/release-pack";
@@ -139,4 +140,5 @@ export const TOOL_PACK_REGISTRY = composeToolPacks([
   decompositionPack,
   buildVisibilityPack,
   buildEvidenceExtraPack,
+  principleDecidePack,
 ]);
